Cython search_spaces v1

+++ xiangqigame/starting_board_builder.py
@@ -36,7 +36,7 @@ class StartingBoardBuilder:
         return black_half + red_half
 
     def build_initial_board_array(self):
-        return np.array(self.build_initial_board_list())
+        return np.array(self.build_initial_board_list(), dtype=np.intc)
 
 
 if __name__ == "__main__":
